Added a simple texture manager
Also restructured the project files
